refactor: rearrange poetry commands component
