+++ components/polylith/bricks/__init__.py
@@ -1,10 +1,8 @@
 from polylith.bricks.base import create_base, get_bases_data
-from polylith.bricks.brick import create_brick
 from polylith.bricks.component import create_component, get_components_data
 
 __all__ = [
     "create_base",
-    "create_brick",
     "create_component",
     "get_bases_data",
     "get_components_data",
