+++ components/polylith/bricks/base.py
@@ -1,12 +1,13 @@
 from pathlib import Path
 
-from polylith.bricks import brick, component
+from polylith.bricks import component
+from polylith.bricks.brick import create_brick
 from polylith.repo import bases_dir
 from polylith.test import create_test
 
 
 def create_base(path: Path, namespace: str, package: str) -> None:
-    brick.create_brick(path, bases_dir, namespace, package)
+    create_brick(path, bases_dir, namespace, package)
     create_test(path, bases_dir, namespace, package)
