+++ development/david.py
@@ -1,14 +1,12 @@
 from polylith import (
     poetry_plugin,
-    base,
     bricks,
-    component,
     development,
     dirs,
     files,
     interface,
     log,
-    poetry_command,
+    poetry,
     project,
     readme,
     repo,
