+++ pyproject.toml
@@ -8,15 +8,13 @@ repository = "https://github.com/davidvujic/python-polylith"
 readme = "README.md"
 packages = [
          {include = "polylith/poetry_plugin", from = "bases"},
-         {include = "polylith/base", from = "components"},
-         {include = "polylith/block", from = "components"},
-         {include = "polylith/component", from = "components"},
+         {include = "polylith/bricks", from = "components"},
          {include = "polylith/development", from = "components"},
          {include = "polylith/dirs", from = "components"},
          {include = "polylith/files", from = "components"},
          {include = "polylith/interface", from = "components"},
          {include = "polylith/log", from = "components"},
-         {include = "polylith/poetry_command", from = "components"},
+         {include = "polylith/poetry", from = "components"},
          {include = "polylith/project", from = "components"},
          {include = "polylith/readme", from = "components"},
          {include = "polylith/repo", from = "components"},
